refactor: Create common helper to check dialect relevance.
