+++ zql/grammar_test.py
@@ -1,4 +1,4 @@
-from zql.grammar import parse_grammar
+from zql.grammar import parse_grammar, is_relevant_to_dialect
 from zql.sample_grammars import (
     ENGLISH_TRANSLATION_GRAMMAR_CONTENT,
     FORMULA_GRAMMAR_CONTENT,
@@ -106,3 +106,19 @@ def test_parse_grammar_english():
         ],
     }
     assert actual == expected
+
+
+def test_is_relevant_to_dialect():
+    # Object supports any dialect
+    assert is_relevant_to_dialect({"dialects": []}, None)
+    assert is_relevant_to_dialect({}, None)
+    assert is_relevant_to_dialect({"dialects": []}, "a")
+    assert is_relevant_to_dialect({}, "a")
+    # Default dialect requested, but object only supports specific dialects
+    assert not is_relevant_to_dialect({"dialects": ["a"]}, None)
+    # Specific dialect requested and object supports it
+    assert is_relevant_to_dialect({"dialects": ["a"]}, "a")
+    assert is_relevant_to_dialect({"dialects": ["a", "b"]}, "a")
+    assert is_relevant_to_dialect({"dialects": ["a", "b"]}, "b")
+    # Specific dialect requested, but object does not support it
+    assert not is_relevant_to_dialect({"dialects": ["a", "b"]}, "z")
